Move applyLanguage into its own file
