+++ src/redux/sagas/__tests__/preferences.test.tsx
@@ -8,16 +8,11 @@
  * This file is copyright under the latest version of the EUPL.
  * Please see LICENSE file for your rights under this license. */
 
-import i18n from "i18next";
 import { call } from "redux-saga/effects";
 import { PayloadAction } from "redux-starter-kit";
 import { preferencesSuccess } from "../../actions";
-import {
-  applyLanguage,
-  cachePreferences,
-  fetchPreferences
-} from "../preferences";
-import { expectSaga, testSaga } from "redux-saga-test-plan";
+import { cachePreferences, fetchPreferences } from "../preferences";
+import { expectSaga } from "redux-saga-test-plan";
 import api from "../../../util/api";
 import config from "../../../config";
 import { WEB_PREFERENCES_STORAGE_KEY } from "../../state/preferences";
@@ -61,25 +56,5 @@ describe("cachePreferences", () => {
         JSON.stringify(preferences)
       )
       .run();
-  });
-});
-
-describe("applyLanguage", () => {
-  it("should not update i18next if the language is the same", () => {
-    i18n.language = "test-language";
-
-    testSaga(applyLanguage, action)
-      .next()
-      .isDone();
-  });
-
-  it("should update i18next if the language is different", () => {
-    i18n.language = "en";
-
-    testSaga(applyLanguage, action)
-      .next()
-      .call([i18n, "changeLanguage"], "test-language")
-      .next()
-      .isDone();
   });
 });
